+++ src/index.ts
@@ -6,7 +6,6 @@ export * from './chip'
 export * from './divider'
 // export * from './dropzone'
 export * from './icon'
-export * from './layout'
 export * from './menu'
 export * from './modal'
 export * from './radio'
