+++ tsconfig.json
@@ -19,6 +19,6 @@
     "emitDeclarationOnly": true,
     "jsxImportSource": "@emotion/react"
   },
-  "include": ["src", "types"],
+  "include": ["./src", "./types"],
   "exclude": ["./node_modules", "./dist", "./src/**/*.stories.tsx"]
 }
